pyproject.toml
[tool.pdm]

[project]
name = "depwatch"
version = "0.1.0"
description = ""
authors = [
    {name = "hamakou108", email = "hamakou108@gmail.com"},
]
dependencies = [
    "typer[all]>=0.7.0",
    "PyGithub>=1.58.0",
    "python-dotenv>=0.21.1",
    "pycircleci>=0.6.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.2.1",
    "pytest-mock>=3.10.0",
]
linter = [
    "mypy>=1.0.1",
    "flake8>=6.0.0",
    "Flake8-pyproject>=1.2.2",
    "black>=23.1.0",
]

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[[tool.mypy.overrides]]
module = "pycircleci.*"
ignore_missing_imports = true

[tool.flake8]
select = ['F']

[tool.black]
line-length = 88
